requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
idna==3.4
IMAPClient==2.3.1
pycparser==2.21
pyspnego==0.10.2
python-dotenv==1.0.0
pywinrm==0.4.3
requests==2.31.0
requests-ntlm==1.2.0
six==1.16.0
smbprotocol==1.11.0
urllib3==2.0.7
xmltodict==0.13.0
yadisk==1.3.4
